Tweaked code that disables Firefox's validation
This change allows me to disable the red outline and tooltip that
Firefox puts on the inputs if the value doesn’t match the pattern,
without the added complexity of a form element within the container. It
is kind of hacky but I think it is the best option.

Removing the pattern attribute from the inputs will disable the
validation but it will make iOS show the text keyboard instead of the
number keyboard.

Using number inputs instead of text inputs will show the number
keyboard, but it will also show ugly arrow buttons that can’t be
removed in any cross-browser way AFAIK.