Added function to solve sudoku puzzles
This may not be exposed to the user, but it may help to generate
puzzles that are solvable, but with predictable difficulty, possibly by
solving a blank grid and then removing some digits. I borrowed heavily
from Peter Norvig’s techniques, involving constraint propagation and
depth-first search, at http://norvig.com/sudoku.html but the code is
all mine.