Fixed bugs with arrow keys sometimes not working
I wanted the left and right arrow keys to move from one character to
another, as usual, unless the cursor was at the beginning or end of the
value in the input, in which case it would move to the next or previous
cell.

This is harder than I thought. Recent versions of Chrome and Firefox
don’t have a way to get the cursor position in a number input,
following the WHAT-WG’s mysterious recommendation
(http://www.whatwg.org/specs/web-apps/current-work/multipage/the-input-e
lement.html#do-not-apply). Trying to get the cursor position in Firefox
throws weird exceptions and prevents the rest of my code from working.

I could use text inputs instead, but that would make it tricky to get
the number keyboard to show up in iOS, and impossible, as far as I can
tell, to make it show up in Android.

I decided to use tel inputs because Chrome and Firefox allow me to get
the cursor position, and Android and iOS show something resembling
their simple number keyboard. I think this is the best solution, unless
I can think of something better.